Separate test file
The test inputs have very long (3363 columns) lines.
